Finish Learning Session Feature api

+++ WeLearnApp/src/main/java/com/welearn/WeLearnApp/entity/LearningSession.java
@@ -63,5 +63,5 @@ public class LearningSession {
     UserProfile tutor;
 
     @Column(name = "tuition")
-    int tuition;
+    double tuition;
 }
